# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "Example info",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://example.localhost",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/notifications": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Create a new notification",
                "description": "Creates a new notification.",
                "operationId": "6c58a9d6ff24e266a1fed03f114f66fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_user_creator",
                                    "id_user_receiver",
                                    "text"
                                ],
                                "properties": {
                                    "id_user_creator": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_user_receiver": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Task completed"
                                    },
                                    "id_task_progress": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "task_status_change"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notification created"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            }
        },
...
        "/troops/{id_troop}/patrols/{id_patrol}/members": {
            "get": {
                "tags": [
                    "Troops",
                    "Patrols"
                ],
                "summary": "Get all patrol members within a troop",
                "description": "Retrieves all members of a specific patrol belonging to a troop.",
                "operationId": "0b1bc57eda4a1f33443159d129c25824",
                "parameters": [
                    {
                        "name": "id_troop",
                        "in": "path",
                        "description": "ID of the troop",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_patrol",
                        "in": "path",
                        "description": "ID of the patrol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of patrol members"
                    },
                    "403": {
                        "description": "Patrol does not belong to troop"
                    },
                    "404": {
                        "description": "Patrol not found"
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            }
        },
        "/patrol/check-invite": {
            "post": {
                "tags": [
                    "Patrols"
                ],
                "summary": "Check if invite code is valid",
                "description": "Validates whether the given invite code corresponds to an existing patrol.",
                "operationId": "142aac17a2e9387218d4d643826eb806",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "invite_code"
                                ],
                                "properties": {
                                    "invite_code": {
                                        "type": "string",
                                        "example": "abc123xyz"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Valid invite code, patrol found"
                    },
                    "400": {
                        "description": "Missing or invalid invite code"
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            }
        },
        "/patrols/{id_patrol}/leaders": {
            "post": {
                "tags": [
                    "Patrols"
                ],
                "summary": "Add a patrol leader to a patrol",
                "description": "Creates a new patrol leader in the specified patrol.",
                "operationId": "a04076fe89a611022c07f7256c607ece",
                "parameters": [
                    {
                        "name": "id_patrol",
                        "in": "path",
                        "description": "ID of the patrol to assign leader to",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_user"
                                ],
                                "properties": {
                                    "id_user": {
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patrol leader assigned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GangLeader"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input or patrol ID"
                    },
                    "500": {
                        "description": "Database error"
                    }
                }
            }
        },
        "/users/{id_user}/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get all notifications for a user",
                "description": "Retrieves all notifications received by the specified user.",
                "operationId": "86a78643707bd214ca9568eb7f32908f",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of notifications"
                    }
                }
            }
        },
        "/notifications/{id_notification}": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Update a notification",
                "description": "Updates a notification (e.g. marking it as received).",
                "operationId": "77186024dffcd21548f66c7b3b14cc2f",
                "parameters": [
                    {
                        "name": "id_notification",
                        "in": "path",
                        "description": "ID of the notification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "was_received": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification updated"
                    },
                    "404": {
                        "description": "Notification not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GangLeader": {
                "title": "GangLeader",
                "description": "Represents a patrol leader (gang leader) within a patrol",
                "properties": {
                    "id_patrol_leader": {
                        "description": "GangLeader model representing a user's role as a leader of a patrol (gang).\n\nIncludes patrol and troop metadata for use in enriched API responses.\n\n/**",
                        "type": "integer",
                        "example": 12
                    },
                    "id_user": {
                        "type": "integer",
                        "example": 123
                    },
                    "id_patrol": {
                        "type": "integer",
                        "example": 5
                    },
                    "patrol_name": {
                        "type": "string",
                        "example": "Foxes"
                    },
                    "patrol_color": {
                        "type": "string",
                        "example": "#ff6600"
                    },
                    "id_troop": {
                        "type": "integer",
                        "example": 2
                    },
                    "troop_name": {
                        "type": "string",
                        "example": "Northwind Troop"
                    }
                },
                "type": "object"
            }
        }
    },
...
}